+++ config/urls.py
@@ -21,6 +21,8 @@ urlpatterns = [
     path("join/", views.join_index.as_view(), name="join_index"),
     path("list/", views.list_index.as_view(), name="list_index"),
     path('gameday/<uuid:gameday_id>/', views.game, name='game'),
+    path('gamedayedit/<uuid:gameday_id>/', views.gameedit, name='gameedit'),
+
     path('gameday_csv/<uuid:gameday_id>/', views.create_csv_gameday, name='gameday_csv'),
 
     path("gamedays/", views.gamedays_index.as_view(), name="gamedays_index"),