+++ config/views.py
@@ -100,8 +100,58 @@ def game(request, gameday_id, sort_field="player__name", state_filter=-1):
     html += (
         f"<a href=/gameday/{gameday.id}/?sort=state&filter={state_filter}>Status</a></h2>  "
     )
+    html += f"<h2><a href=/gamedayedit/{gameday.id}>Edit Player states</a></h2>"
 
     template = loader.get_template("pages/game.html")
+    context = {
+        'games': games,
+        'gameday': gameday,
+        'availability': avail,
+        'whatsapp_text':whatsapp_text,
+        'totals' : html
+    }
+    return HttpResponse(template.render(context, request))
+
+def gameedit(request, gameday_id, sort_field="player__name", state_filter=-1):
+    gameday = Gameday.objects.get(id=gameday_id)
+    games = gameday.games.all()
+    whatsapp_text=f"?text=Are%20you%20available%20for%20{gameday.date}?%20Please%20update%20Waatea!"
+
+    dontknow, no, yes, notset = calc_avail_totals(gameday)
+    html = ""
+    html += "<p>"
+    html += f'<h2><div style="color:black">Not set: <a href="/gamedayedit/{gameday.id}/?filter=0">{notset}</a></div></h2>'
+    html += f'<h2><div style="color:orange">Not sure: <a href="/gamedayedit/{gameday.id}/?filter=1">{dontknow}</a></div></h2>'
+    html += f'<h2><div style="color:red">No: <a href="/gamedayedit/{gameday.id}/?filter=2">{no}</a></div></h2>'
+    html += f'<h2><div style="color:green">Yes!: <a href="/gamedayedit/{gameday.id}/?filter=3"> {yes}</a></div></h2>'
+    html += "</p>"
+
+    if request.method == "GET":
+        sort_field = request.GET.get("sort", sort_field)
+        state_filter = int(request.GET.get("filter", state_filter))
+
+    if 0 <= state_filter <= 4:
+        avail = Availbility.objects.filter(
+            gameday=gameday, player__active=True, state=state_filter
+        ).order_by(sort_field)
+        state_name = Availbility.STATE_CHOICES[state_filter][-1]
+        html += f'<h2><div style="color:black">Showing only state {state_name}. '
+        html += f'<a href="/gamedayedit/{gameday.id}">Click to remove</a></div></h2>'
+    else:
+        avail = Availbility.objects.filter(
+            gameday=gameday, player__active=True
+        ).order_by(sort_field)
+
+    update_availlist(gameday)
+    html += "<p>"
+    html += "<h2>Sort by: "
+    html += f"<a href=/gamedayedit/{gameday.id}/?sort=player__name&filter={state_filter}>Player</a>,  "
+    html += f"<a href=/gamedayedit/{gameday.id}/?sort=-updated&filter={state_filter}>Last Updated</a>,  "
+    html += (
+        f"<a href=/gamedayedit/{gameday.id}/?sort=state&filter={state_filter}>Status</a></h2>  "
+    )
+
+    template = loader.get_template("pages/gameedit.html")
     context = {
         'games': games,
         'gameday': gameday,